Solution to Calculate and Print the Average of Three Numbers

+++ Calculate and Print the Average of Three Numbers@@6788d5012db25efec946f944/code.c
@@ -3,7 +3,7 @@ int main(){
     int a,b,c;
     float avergae;
     scanf("%d %d %d",&a , &b, &c);
-    avergae= (a+b+c)/3;
+    avergae= (float)(a+b+c)/3;
     printf("Average: %.2f",avergae);
 }
